The feasibility functions now return FALSE instead of an error for primaldualunknown.
For this some duplicated code was needed as for example primalfeasible cannot just return !primalinfeasible, because then it would return true for pdunkown.
This was done so that the code won't just stop if this happens but continues with penalty-formulation or in the worst case copies the bound of the parent node.