expanded sdpobjective branching rule to check coupled variables if candidates objective is 0
If the objective of all candidates is 0, the branching rule will now get all variables, that are
directly coupled with one of the candidates via a constraint, and combine the absolute 
objectives of all such variables to get the candidates total objective, which is then used
to decide on which variable to branch.

This could be expanded to remove all variables, that are coupled to multiple candidates
and only take those into account, that are only coupled to a single candidate.

Also added the SCIPconsGetVars and NVars functionality to the SDP Constraint Handler,
as this is used in the branching rule.