added another branching and node selection rule that tries to approximate the CS-solution
The new solution after fixing a binary variable to zero is calculated by setting all coresponding
matrix entries to zero and scaling the rest by 1/1-x, where x is the corresponding diagonal
entry.
The branching is then conducted on the variable, for which this value is the smallest, this
value is also used for node selection for the node where the branching candidate is fixed to
zero, for the node with fixation to one, the old objective is used.

For the node selection a new node selector was implemented, that just chooses the node
with the highest priority, this had to be done via a hashmap, as the node priorities apparently
aren't accessable in SCIP.