added possibility to also check slater condition for primal problem, currently deactivated
unfortunately this check depends on the slater condition of a modified problem with removed
constant part (and one added LP-constraint), as this problem will be solved to get the biggest
possible eigenvalue as the optimal solution of the primal problem (we can't solve it directly
because of the equality constraints). Unfortunately the slater condition for this modified
example doesn't hold for basically all applications.

also made some corrections to the dual test for the case that the dual problem is unbounded